Fiz o componente principal e ajeitei alguns problemas
