+++ whatslab/src/componentes/JanelaDeChat/JanelaDeChat.css
@@ -1,0 +1,27 @@
+#form {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .fala {
+    font-weight: bold;
+  }
+  
+  #nome {
+    width: 120px;
+  }
+  
+  #texto {
+    width: 220px;
+    margin-left: 5px;
+    margin-right: 5px;
+  }
+  
+  #main {
+    padding: 10px;
+    border: 2px solid black;
+    height: 75vh;
+    width: 50vw;
+    margin-left: 25vw;
+  }
